feat: flatten notify quiet support
